쿠파 점프 구현
.

+++ finalProject/finalProject/boss.h
@@ -27,7 +27,7 @@ private:
 
 	int frameX;
 	int frameY;
-	int jumpnum;
+	float jumpnum;
 	int frameNum;
 
 	POINT* targetPoint;
